Removing omm-util lib and moving it in to omm-lib
