+++ libraries/omm-lib/src/log.h
@@ -12,7 +12,7 @@ license agreement from NVIDIA CORPORATION is strictly prohibited.
 
 #include "defines.h"
 #include <omm.h>
-#include <shared/assert.h>
+#include "util/assert.h"
 
 #include <cstdio>
