+++ libraries/omm-lib/src/texture_impl.cpp
@@ -13,9 +13,9 @@ license agreement from NVIDIA CORPORATION is strictly prohibited.
 #include "defines.h"
 #include "std_containers.h"
 
-#include <shared/math.h>
-#include <shared/bit_tricks.h>
-#include <shared/texture.h>
+#include "util/math.h"
+#include "util/bit_tricks.h"
+#include "util/texture.h"
 
 #include <cstring>
