+++ libraries/omm-lib/src/util/bird.h
