+++ libraries/omm-lib/src/util/bit_tricks.h
