+++ libraries/omm-lib/src/util/geometry.h
