+++ libraries/omm-lib/src/util/math.h
