+++ libraries/omm-lib/src/util/util.h
